+++ projects/js-packages/publicize-components/jest-globals.js
@@ -1,0 +1,12 @@
+if ( ! window.matchMedia ) {
+	window.matchMedia = query => ( {
+		matches: false,
+		media: query,
+		onchange: null,
+		addListener: jest.fn(), // deprecated
+		removeListener: jest.fn(), // deprecated
+		addEventListener: jest.fn(),
+		removeEventListener: jest.fn(),
+		dispatchEvent: jest.fn(),
+	} );
+}