+++ projects/packages/videopress/src/client/admin/components/video-card/types.ts
@@ -1,114 +1,31 @@
-import { MouseEvent } from 'react';
+import { VideoPressVideo } from '../../types';
+import { VideoQuickActionsProps } from '../video-quick-actions/types';
+import { VideoThumbnailProps } from '../video-thumbnail/types';
 
-export type VideoThumbnailProps = {
-	/**
-	 * className to apply to the component
-	 */
-	className?: string;
+export type VideoCardProps = Pick< VideoPressVideo, 'title' | 'plays' | 'id' > &
+	VideoThumbnailProps &
+	VideoQuickActionsProps & {
+		/**
+		 * Should show or not quick actions.
+		 */
+		showQuickActions?: boolean;
+		/**
+		 * Callback to be invoked when clicking on the `Edit video details` button.
+		 */
+		onVideoDetailsClick?: () => void;
 
-	/**
-	 * Video thumbnial image
-	 */
-	thumbnail: string;
+		/**
+		 * True when the video is being deleted.
+		 */
+		isDeleting?: boolean;
 
-	/**
-	 * Video duration. Number, in milliseconds.
-	 */
-	duration?: number;
+		/**
+		 * True when is in loading mode.
+		 */
+		loading?: boolean;
 
-	/**
-	 * Whether is possible to edit the thumbnail
-	 */
-	editable?: boolean;
-};
-
-export type VideoThumbnailDropdownProps = {
-	/**
-	 * Callback to be invoked when clicking on the `Use default thumbnail` dropdown menu option.
-	 */
-	onUseDefaultThumbnail?: () => void;
-
-	/**
-	 * Callback to be invoked when clicking on the `Select from video` dropdown menu option.
-	 */
-	onSelectFromVideo?: () => void;
-
-	/**
-	 * Callback to be invoked when clicking on the `Upload image` dropdown menu option.
-	 */
-	onUploadImage?: () => void;
-};
-
-export type VideoDetailsProps = {
-	/**
-	 * Video filename.
-	 */
-	filename: string;
-
-	/**
-	 * Video source file URL.
-	 */
-	src: string;
-
-	/**
-	 * Video uploaded date
-	 */
-	uploadDate: string;
-};
-
-export type VideoCardProps = {
-	/**
-	 * Video ID
-	 */
-	id: number | string;
-
-	/**
-	 * Video title
-	 */
-	title: string;
-
-	/**
-	 * Video uploaded date
-	 */
-	uploadDate: string;
-
-	/**
-	 * Video thumbnial/poster image URL
-	 */
-	posterImage?: string;
-
-	/**
-	 * Video duration, in milliseconds
-	 */
-	duration?: number;
-
-	/**
-	 * Plays counter
-	 */
-	plays?: number;
-
-	/**
-	 * Whether the video is private, or not.
-	 */
-	isPrivate?: boolean;
-
-	/**
-	 * Callback to be invoked when clicking on the `Edit video details` button.
-	 */
-	onVideoDetailsClick?: () => void;
-
-	/**
-	 * Callback to be invoked when clicking on the `Update thumbnail` button.
-	 */
-	onUpdateThumbnailClick?: ( event: MouseEvent< HTMLButtonElement > ) => void;
-
-	/**
-	 * Callback to be invoked when clicking on the `Update privacy` button.
-	 */
-	onUpdateUpdatePrivacyClick?: ( event: MouseEvent< HTMLButtonElement > ) => void;
-
-	/**
-	 * Callback to be invoked when clicking on the `Delete video` button.
-	 */
-	onDeleteClick?: ( event: MouseEvent< HTMLButtonElement > ) => void;
-};
+		/**
+		 * True when the video has been deleted in the server-side.
+		 */
+		hasBeenDeleted?: boolean;
+	};