+++ projects/packages/videopress/src/client/admin/index.js
@@ -1,7 +1,24 @@
 import { ThemeProvider } from '@automattic/jetpack-components';
-import React from 'react';
 import ReactDOM from 'react-dom';
+import { HashRouter, Routes, Route } from 'react-router-dom';
+import { initStore } from '../state';
 import AdminPage from './components/admin-page';
+import EditVideoDetails from './components/edit-video-details';
+
+initStore();
+
+const VideoPress = () => (
+	<ThemeProvider>
+		<HashRouter>
+			<Routes>
+				<Route path="/" element={ <AdminPage /> } />
+				<Route path="video">
+					<Route path=":videoId/edit" element={ <EditVideoDetails /> } />
+				</Route>
+			</Routes>
+		</HashRouter>
+	</ThemeProvider>
+);
 
 /**
  * Initial render function.
@@ -13,12 +30,7 @@ function render() {
 		return;
 	}
 
-	ReactDOM.render(
-		<ThemeProvider>
-			<AdminPage />
-		</ThemeProvider>,
-		container
-	);
+	ReactDOM.render( <VideoPress />, container );
 }
 
 render();