+++ projects/plugins/social/tests/php/test-class-jetpack-social.php
@@ -6,7 +6,6 @@
  */
 
 use Automattic\Jetpack\Connection\Manager as Connection_Manager;
-use Automattic\Jetpack\Modules;
 use WorDBless\BaseTestCase;
 
 /**
@@ -38,10 +37,10 @@ class Jetpack_Social_Test extends BaseTestCase {
 	 */
 	public function test_publicize_module_is_activated_on_plugin_activation() {
 		$this->activate_plugin( 'hello-world.php' );
-		$this->assertFalse( ( new Modules() )->is_active( Jetpack_Social::JETPACK_PUBLICIZE_MODULE_SLUG ) );
+		$this->assertFalse( ( Jetpack_Social::is_publicize_active() ) );
 
 		$this->activate_plugin( JETPACK_SOCIAL_PLUGIN_ROOT_FILE_RELATIVE_PATH );
-		$this->assertFalse( ( new Modules() )->is_active( Jetpack_Social::JETPACK_PUBLICIZE_MODULE_SLUG ) );
+		$this->assertFalse( ( Jetpack_Social::is_publicize_active() ) );
 
 		$connection_manager = $this->createMock( Connection_Manager::class );
 		$connection_manager->method( 'is_connected' )->willReturn( true );
@@ -54,7 +53,7 @@ class Jetpack_Social_Test extends BaseTestCase {
 		$this->social->expects( $this->once() )->method( 'calculate_scheduled_shares' );
 
 		$this->activate_plugin( JETPACK_SOCIAL_PLUGIN_ROOT_FILE_RELATIVE_PATH );
-		$this->assertTrue( ( new Modules() )->is_active( Jetpack_Social::JETPACK_PUBLICIZE_MODULE_SLUG ) );
+		$this->assertTrue( Jetpack_Social::is_publicize_active() );
 
 	}
 }