+++ projects/packages/waf/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.7] - 2022-09-20
+### Changed
+- Changing how we load and run the package to avoid actions.php [#24730]
+
+## [0.6.6] - 2022-09-08
+### Fixed
+- Fixed exception namespace. [#25663]
+
 ## [0.6.5] - 2022-07-26
 ### Changed
 - Updated package dependencies. [#25158]
@@ -84,6 +92,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Core: do not ship .phpcs.dir.xml in production builds.
 
+[0.6.7]: https://github.com/Automattic/jetpack-waf/compare/v0.6.6...v0.6.7
+[0.6.6]: https://github.com/Automattic/jetpack-waf/compare/v0.6.5...v0.6.6
 [0.6.5]: https://github.com/Automattic/jetpack-waf/compare/v0.6.4...v0.6.5
 [0.6.4]: https://github.com/Automattic/jetpack-waf/compare/v0.6.3...v0.6.4
 [0.6.3]: https://github.com/Automattic/jetpack-waf/compare/v0.6.2...v0.6.3